+++ php/Terminus/Commands/SiteCommand.php
@@ -1305,8 +1305,8 @@ class SiteCommand extends TerminusCommand {
    *
    * ## OPTIONS
    *
-   * <clear>
-   * : clear - Clear redis cache on remote server
+   * <enable|disable|clear>
+   * : Options are enable, disable, and clear.
    *
    * [--site=<site>]
    * : site name
@@ -1317,22 +1317,41 @@ class SiteCommand extends TerminusCommand {
    * ## Examples
    *
    *    terminus site redis clear --site=behat-tests --env=live
+   *    terminus site redis enable --site=behat-tests
+   *    terminus site redis disable --site=behat-tests
    */
   public function redis($args, $assoc_args) {
     $action = array_shift($args);
     $site   = $this->sites->get(Input::siteName(['args' => $assoc_args]));
-    if (isset($assoc_args['env'])) {
-      $environments = [$site->environments->get($assoc_args['env'])];
-    } else {
-      $environments = $site->environments->all();
+    if (in_array($site->info('service_level'), ['free', 'basic', 'pro'])) {
+      $this->failure(
+        'You must upgrade to a business or an elite plan to use Redis.'
+      );
     }
     switch ($action) {
+      case 'enable':
+        $redis = $site->enableRedis();
+        if ($redis) {
+          $this->log()->info('Redis enabled. Converging bindings...');
+        }
+          break;
+      case 'disable':
+        $redis       = $site->disableRedis();
+        if ($redis) {
+          $this->log()->info('Redis disabled. Converging bindings...');
+        }
+          break;
       case 'clear':
+        if (isset($assoc_args['env'])) {
+          $environments = [$site->environments->get($assoc_args['env'])];
+        } else {
+          $environments = $site->environments->all();
+        }
         $commands = array();
         foreach ($environments as $environment) {
           $connection_info = $environment->connectionInfo();
           if (!isset($connection_info['redis_host'])) {
-            $this->failure('Redis cache not enabled');
+            $this->failure('Redis cache is not enabled.');
           }
           $args = [
             $environment->get('id'),
@@ -1351,11 +1370,11 @@ class SiteCommand extends TerminusCommand {
             }
           );
           $command  = 'ssh -p 2222 %s.%s@appserver.%s.%s.drush.in';
-          $command .= ' "redis-cli -h %s -p %s -a %s flushall"';
+          $command .= ' "redis-cli -h %s -p %s -a %s flushall" 2> /dev/null';
           $commands[$environment->get('id')] = vsprintf($command, $args);
         }
         foreach ($commands as $env => $command) {
-          $this->log()->info('Clearing redis on {env}', array('env' => $env));
+          $this->log()->info('Clearing Redis on {env}.', compact('env'));
           exec($command, $stdout, $return);
           if (!empty($stdout)) {
             $this->log()->info($stdout[0]);